add if statement that checks array elements are not empty
In the case that we run Gaussian and scrub any points, the following AIMALL calculation still has the same number of tasks as the Gaussian job, but the AIMALL datafile will have less files (because points are scrubbed). This will mean the last couple of tasks in the AIMALL script will fail because array elements will have empty entries (as there are not less points, thus less files to run through AIMALL). Failing the last tasks writes out error files to the ERRORS directory, which say that there is an "ambiguous redirect" because the array value that is being accessed is empty. This also makes the checkmanager for aimall raise an error as the Path defaults to "." (because an empty path was passed in initially), which raises an additional error. This bug DOES NOT cause calculations to fail or write out information to the wrong places, so technically not harmful, but still better to fix.