always use wfn file for geometry
Always try to use wfn file for geometry initially. This is because the .wfn file contains the final geometry. If gaussian optimizes, this final geometry will be different from the initial .xyz/.gjf geometry. Therefore, when doing any predictions on this, they will be way off because you could be using .gjf/.xyz geometry for predictions, but getting true values which were calculated for the .wfn geometry