add try/except clause
Added a try except clause which handles if a point does not have .int files for some reason. If the point does not have int files, then point[atom] (which is an AtomData instance) does not have the .properties attribute set (default is None). The properties attribute should be an instance of class INT (which then subclasses from GeometryDataFile where the get_property method is). But if there is no int file, this part of the code will error out. Now it should write out a warning to the log file and make the models using the points which actually have the data in them and ignore the rest.