remove some cyclic imports and fixes
A very large commit, containing formatting, as well as moving a lot of files around because the files before were ordered very poorly.
This should fix some of the cyclic imports as well as make the package look more like a library.
Also a lot of unnecessary stuff relating to the submit of compute were removed because they cluttered the code